Refactored zendesk ticket approval into its own function
